Add threaded publisher code

+++ src/main/java/uk/ac/soton/ldanalytics/hubber/bench/Point.java
@@ -2,14 +2,14 @@ package uk.ac.soton.ldanalytics.hubber.bench;
 
 public class Point {
 	private String msg;
-	private Long publisher;
+	private int publisher;
     
-    public void set(Long publisher, String msg) {
+    public void set(int publisher, String msg) {
     	this.publisher = publisher;
         this.msg = msg;
     }
     
-    public long getPublisher() {
+    public int getPublisher() {
     	return publisher;
     }
